+++ metrics-guice/src/main/java/com/yammer/metrics/guice/GaugeInjectionListener.java
@@ -11,13 +11,13 @@ import java.lang.reflect.Method;
  * An injection listener which creates a gauge for the declaring class with the given name (or the
  * method's name, if none was provided) which returns the value returned by the annotated method.
  */
-public class GaugeInjectionListener<I> implements InjectionListener<I> {
+class GaugeInjectionListener<I> implements InjectionListener<I> {
     private final MetricsRegistry metricsRegistry;
     private final TypeLiteral<I> literal;
     private final String name;
     private final Method method;
 
-    public GaugeInjectionListener(MetricsRegistry metricsRegistry, TypeLiteral<I> literal, String name, Method method) {
+    GaugeInjectionListener(MetricsRegistry metricsRegistry, TypeLiteral<I> literal, String name, Method method) {
         this.metricsRegistry = metricsRegistry;
         this.literal = literal;
         this.name = name;